requirements.txt
# frappe -- https://github.com/frappe/frappe is installed via 'bench init'
qrcode[pil]>=7.3.1
Pillow>=9.0.0
